requirements.txt
absl-py==2.0.0
anyio==4.0.0
appnope==0.1.3
astor==0.8.1
asttokens==2.4.0
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
comm==0.1.4
contourpy==1.1.1
crypto==1.4.1
cryptography==38.0.4
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.13
et-xmlfile==1.1.0
executing==2.0.0
fastjsonschema==2.18.1
firebase==3.0.1
flatbuffers==23.5.26
fonttools==4.43.1
gast==0.5.4
gcloud==0.18.3
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.57.0
grpcio==1.59.0
h11==0.14.0
h5py==3.9.0
httpcore==0.18.0
httplib2==0.21.0
httpx==0.25.0
idna==3.4
imageio==2.31.6
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.4.0
jupyter_core==5.4.0
jwcrypto==1.4.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
Naked==0.1.32
nbformat==5.9.2
nest-asyncio==1.5.8
numpy==1.26.0
oauth2client==4.1.3
oauthlib==3.2.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybase64==1.3.1
pycparser==2.21
pycryptodome==3.16.0
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2023.3.post1
PyYAML==6.0
pyzmq==25.1.1
referencing==0.30.2
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rpds-py==0.10.6
rsa==4.9
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
sseclient==0.0.27
stack-data==0.6.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-docs @ git+https://github.com/tensorflow/docs@159143b05575c90bc1217bfa99c76cb5ca27b8d4
tensorflow-estimator==2.14.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.8
Werkzeug==3.0.0
wrapt==1.14.1
